fix units

+++ models/intermediate/intermediate__track_stats.sql
@@ -11,7 +11,7 @@ select
     min(sh.played_at) as first_played_at,
     max(sh.played_at) as last_played_at,
     coalesce(sum(is_skipped::int), 0) as times_skipped,
-    coalesce(sum(is_skipped::int) / count(sh.track_id), 0) as skip_rate,
+    coalesce(sum(is_skipped::int)::float / count(sh.track_id), 0) as skip_rate,
     coalesce(count(*) filter (where reason_start = 'clickrow'), 0) as manual_plays
 from streaming_history sh
 group by sh.track_id
